Fixes potions tooltip, adds abstract box, removes a lot of redundant code
